Simplify the files in the `jni` folder into one and move the resulting file to the parent folder
